pyproject.toml
[tool.poetry]
name = "markhub"
version = "0.1.0"
description = "Markdown content development service for your GitHub repositories"
authors = ["Roman Yatsenko  <yatsenkoroma@gmail.com>"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = "^3.8"
Django = "~3.2"
django-allauth = "~0.45"
PyGithub = "~1.55"
django-environ = "~0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
